when moving between files, any changed tags will be highlighted
